Get SMS working, make TS happy

+++ database/migrations/2023_07_14_113612_create_listings_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->foreignUuid('seller_id')->nullable()->constrained('users', 'id')->nullOnDelete();
             $table->foreignUuid('buyer_id')->nullable()->constrained('users', 'id')->nullOnDelete();
             $table->string('title', 255)->index();
-            $table->unsignedTinyInteger('status')->default(ListingStatus::AVAILABLE->value)->index();
+            $table->unsignedSmallInteger('status')->default(ListingStatus::AVAILABLE->value)->index();
             $table->bigInteger('price')->nullable();
             $table->dateTime('sold_at')->nullable();
             $table->text('description')->nullable();
